pyproject.toml
[tool.poetry]
name = "undep"
version = "0.1.0"
description = "Indirect dependency management system"
authors = ["Akshit Tyagi <tyagiakshit833@gmail.com>"]
packages = [{include = "undep", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0"
GitPython = "^3.1.0"
fastapi = "^0.104.0"
httpx = "^0.24.0"
pydantic = "^2.0"
rich = "^13.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
black = "^23.0"
isort = "^5.12.0"
mypy = "^1.5.0"

[tool.poetry.scripts]
undep = "undep.cli.main:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"